pyproject.toml
[tool.poetry]
name = "obsipub"
version = "0.1.0"
description = ""
authors = ["rishikanthc <r3chandr@ucsd.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
markdown = "^3.6"
python-frontmatter = "^1.1.0"
fire = "^0.6.0"
loguru = "^0.7.2"
falcon = "^3.1.3"
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
basedpyright = "*"
pyright = "*"
ruff = "*"
pre-commit = "*"
pytest = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "/Users/rishi/miniconda3/envs"
venv = "obsipub"
useLibraryCodeForTypes = true
typeCheckingMode = "standard"
reportUnusedImport = "warning"
reportUnusedVariable = "warning"
reportDuplicateImport = "error"
reportPrivateUsage = "warning"
reportDeprecated = "warning"
reportInconsistentConstructor = "error"
reportMissingSuperCall = "error"
reportUninitializedInstanceVariable = "error"
reportUnnecessaryCast = "warning"
reportUnnecessaryComparison = "warning"
reportImplicitStringConcatenation = "warning"
diagnosticMode = "openFilesOnly"

[tool.poetry.scripts]
html = "scripts.test:main"
